+++ xpeapp_android/app/src/main/java/com/xpeho/xpeapp/data/service/WordpressService.kt
@@ -2,11 +2,13 @@ package com.xpeho.xpeapp.data.service
 
 import com.google.gson.GsonBuilder
 import com.xpeho.xpeapp.BuildConfig
+import com.xpeho.xpeapp.XpeApp
 import com.xpeho.xpeapp.data.entity.AuthentificationBody
 import com.xpeho.xpeapp.data.entity.QvstAnswerBody
 import com.xpeho.xpeapp.data.model.WordpressToken
 import com.xpeho.xpeapp.data.model.qvst.QvstCampaign
 import com.xpeho.xpeapp.data.model.qvst.QvstQuestion
+import com.xpeho.xpeapp.data.service.interceptor.AuthorizationHeaderInterceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
@@ -18,8 +20,13 @@ private const val BASE_URL = BuildConfig.BACKEND_URL
 private val gson = GsonBuilder().setLenient().create()
 
 val logging = HttpLoggingInterceptor().apply { level = HttpLoggingInterceptor.Level.BODY }
+val authorization = AuthorizationHeaderInterceptor(XpeApp.appModule.authenticationManager)
 
-private val okHttpClient = OkHttpClient.Builder().addInterceptor(logging).build()
+private val okHttpClient = OkHttpClient
+    .Builder()
+    .addInterceptor(authorization)
+    .addInterceptor(logging)
+    .build()
 
 private val retrofit =
         Retrofit.Builder()